// client/src/api/generated/models/dto-register-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TaskFuss API
 * API for TaskFuss app
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface DtoRegisterRequest
 */
export interface DtoRegisterRequest {
    /**
     * 
     * @type {string}
     * @memberof DtoRegisterRequest
     */
    'email': string;
    /**
     * 
     * @type {string}
     * @memberof DtoRegisterRequest
     */
    'password': string;
    /**
     * 
     * @type {string}
     * @memberof DtoRegisterRequest
     */
    'username': string;
}