// client/src/api/generated/models/dto-user.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TaskFuss API
 * API for TaskFuss app
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface DtoUser
 */
export interface DtoUser {
    /**
     * 
     * @type {string}
     * @memberof DtoUser
     */
    'created_at'?: string;
    /**
     * 
     * @type {number}
     * @memberof DtoUser
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof DtoUser
     */
    'username'?: string;
}