// client/src/api/generated/models/dto-register-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TaskFuss API
 * API for TaskFuss app
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { DtoUser } from './dto-user';

/**
 * 
 * @export
 * @interface DtoRegisterResponse
 */
export interface DtoRegisterResponse {
    /**
     * 
     * @type {string}
     * @memberof DtoRegisterResponse
     */
    'auth_token'?: string;
    /**
     * 
     * @type {DtoUser}
     * @memberof DtoRegisterResponse
     */
    'user'?: DtoUser;
}