// client/src/api/generated/models/dto-login-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TaskFuss API
 * API for TaskFuss app
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { DtoUser } from './dto-user';

/**
 * 
 * @export
 * @interface DtoLoginResponse
 */
export interface DtoLoginResponse {
    /**
     * 
     * @type {string}
     * @memberof DtoLoginResponse
     */
    'auth_token'?: string;
    /**
     * 
     * @type {DtoUser}
     * @memberof DtoLoginResponse
     */
    'user'?: DtoUser;
}